backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.9.0
psycopg2-binary>=2.9.10
dj-database-url>=3.0.1
python-dotenv>=1.1.1
requests>=2.32.5
